requirements.txt
torch
transformers
accelerate
sentencepiece
openai-whisper
pillow
fastapi
uvicorn[standard]
python-multipart
requests
streamlit
